+++ src/constants/googleUrl.js
@@ -1,0 +1,3 @@
+module.exports = {
+    GOOGLE_REFRESH_URL: 'https://oauth2.googleapis.com/token',
+};
